rename: 폴더명에서 Routes 삭제
