+++ src/routes/Sample/components/SampleComponent.tsx
