+++ src/routes/Sample/styles/SampleComponent.styles.tsx
